bugs on execap=1

+++ scripts/map.rb
@@ -480,7 +480,7 @@ while time.to_f < tstop.to_f
 				cmd = "cp " + PATH + "share/AG_ap.par . "
 				datautils.execute(prefix, cmd);
 				if parameters.execap.to_i == 1
-					binsize = parameters.ranal.to_i * 2
+					binsize = parameters.ranal.to_f * 2
   					cmd = "export PFILES=.:$PFILES; "+PATH+"bin/AG_ap "+listfile+" "+indexlog.to_s+" "+indexfilter.to_s+" "+sarmatrixfull.to_s+" "+edpmatrixfull.to_s+" "+parameters.timelist.to_s+" "+binsize.to_s+" "+parameters.ranal.to_s+" "+l.to_s+" "+b.to_s+" "+lonpole.to_s+" "+" "+parameters.albedorad.to_s+" 0.5 360.0 5.0 "+parameters.phasecode.to_s+" "+parameters.timestep.to_s+" "+parameters.spectralindex.to_s+" "+t0.to_s+" "+t1.to_s+" "+emin.to_s+" "+emax.to_s+" "+fovmin.to_s+" "+fovmax.to_s+" "+parameters.filtercode.to_s+" "+parameters.timeslot.to_s+" 0 "
 				end
 				if parameters.execap.to_i == 2
